backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.7.2
asttokens==2.4.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==5.0.3
django-debug-toolbar==4.3.0
django-redis==5.4.0
djangorestframework==3.15.1
dnspython==2.6.1
eventlet==0.36.1
executing==2.0.1
flower==2.0.1
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==22.0.0
humanize==4.9.0
ipython==8.22.2
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.6
packaging==24.0
parso==0.8.3
pillow==10.2.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
# psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
tornado==6.4
traitlets==5.14.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
